+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Added `GET    /rest/raven/1.0/api/testrun/{id}/status`
 - Added `PUT    /rest/raven/1.0/api/testrun/{id}/status`
 - Added `GET    /rest/raven/1.0/api/testrun/{id}/defect`
+- Added `POST   /rest/raven/1.0/api/testrun/{id}/defect`
+- Added `DELETE /rest/raven/1.0/api/testrun/{id}/defect/{issueIdOrKey}`
 
 # `4.0.0`